requirements.txt
asgiref==3.8.1
Django==5.1.2
django-appconf==1.0.6
django-ckeditor==6.7.1
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.1
tzdata==2024.2
